move UnfoldState to separate file, start a custom flow for flatmapping to it
